# api/openapi.yaml
openapi: 3.0.1
info:
  title: VMClarity APIs
  version: 1.0.0
paths:
...
  /assets:
    get:
      summary: Get assets
      operationId: GetAssets
      parameters:
        - $ref: '#/components/parameters/odataFilter'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataCount'
        - $ref: '#/components/parameters/odataTop'
        - $ref: '#/components/parameters/odataSkip'
        - $ref: '#/components/parameters/odataExpand'
        - $ref: '#/components/parameters/odataOrderBy'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assets'
        default:
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create asset
      operationId: PostAssets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
        required: true
      responses:
        201:
          description: A new asset was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        400:
          description: Invalid asset supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Asset already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExists'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
  /assets/{assetID}:
    get:
      summary: Get asset.
      operationId: GetAssetsAssetID
      parameters:
        - $ref: '#/components/parameters/assetID'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataExpand'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        404:
          description: Asset ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
    put:
      summary: Update asset.
      operationId: PutAssetsAssetID
      parameters:
        - $ref: '#/components/parameters/assetID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
        required: true
      responses:
        200:
          description: Updated asset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        400:
          description: Invalid asset supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Asset ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Asset already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    patch:
      summary: Update asset.
      operationId: PatchAssetsAssetID
      parameters:
        - $ref: '#/components/parameters/assetID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
        required: true
      responses:
        200:
          description: Updated asset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        404:
          description: Asset ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Asset already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    delete:
      summary: Delete asset.
      operationId: DeleteAssetsAssetID
      parameters:
        - $ref: '#/components/parameters/assetID'
      responses:
        200:
          $ref: '#/components/responses/Success'
        400:
          description: Invalid asset supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Asset ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in all cases of failures.
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in cases of success that returns nothing.
    Metadata:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
    MetadataReadOnly:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
          readOnly: true
    Annotations:
      type: array
      description: Generic map of string keys and string values to attach arbitrary non-identifying metadata to objects.
      items:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
    CloudProvider:
      type: string
      enum:
        - AWS
        - Azure
        - GCP
        - Docker
        - External
        - Kubernetes
...
    ScanFindingsSummary:
      description: A summary of the scan findings.
      type: object
      properties:
        totalPackages:
          type: integer
        totalExploits:
          type: integer
        totalMalware:
          type: integer
        totalMisconfigurations:
          type: integer
        totalRootkits:
          type: integer
        totalSecrets:
          type: integer
        totalInfoFinder:
          type: integer
        totalVulnerabilities:
          $ref: '#/components/schemas/VulnerabilityScanSummary'
    VulnerabilityScanSummary:
      description: A summary of number of vulnerabilities found per severity.
      type: object
      properties:
        totalCriticalVulnerabilities:
          type: integer
        totalHighVulnerabilities:
          type: integer
        totalMediumVulnerabilities:
          type: integer
        totalLowVulnerabilities:
          type: integer
        totalNegligibleVulnerabilities:
          type: integer
...
    SecurityGroup:
      type: object
      description: general cloud security group
      properties:
        id:
          type: string
          minLength: 1
      required:
        - id
      additionalProperties: false
    Tag:
      type: object
      description: general cloud tag / label
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
      additionalProperties: false
    RootVolume:
      type: object
      description: Information about VM root volume
      properties:
        sizeGB:
          type: integer
        encrypted:
          type: string
          enum:
            - Yes
            - No
            - Unknown
      required:
        - sizeGB
        - encrypted
      additionalProperties: false
...
    Assets:
      type: object
      properties:
        count:
          type: integer
          description: Total assets count according the given filters
          readOnly: true
        items:
          type: array
          description: List of assets in the given filters and page.
            List length must be lower or equal to pageSize.
          items:
            $ref: '#/components/schemas/Asset'
          readOnly: true
...
    Asset:
      type: object
      description: Describes an asset object.
      allOf:
        - $ref: '#/components/schemas/Metadata'
        - type: object
          properties:
            id:
              type: string
              # TODO(sambetts) Decide if we want the validation here by having
              # separate schemas for GET, POST and PATCH.
              #
              # We can't make anything readOnly or required here if we want to
              # use the same schema for PATCH otherwise it will require you
              # always send the required fields and remove the readOnly fields
              # on the client side.
              #
              # readOnly: true
            revision:
              type: integer
              # TODO(sambetts) Decide if we want the validation here by having
              # separate schemas for GET, POST and PATCH.
              #
              # We can't make anything readOnly or required here if we want to
              # use the same schema for PATCH otherwise it will require you
              # always send the required fields and remove the readOnly fields
              # on the client side.
              #
              # readOnly: true
            assetInfo:
              $ref: '#/components/schemas/AssetType'
            firstSeen:
              type: string
              format: date-time
            terminatedOn:
              type: string
              format: date-time
            lastSeen:
              type: string
              format: date-time
            scansCount:
              description: Total number of scans that have ever run for this asset
              type: integer
            summary:
              $ref: '#/components/schemas/ScanFindingsSummary'
            providers:
              type: array
              description: List of providers which discover the asset.
              items:
                $ref: '#/components/schemas/ProviderRelationship'
...
    AssetExists:
      type: object
      properties:
        message:
          description: Describes which unique constraint combination causes the conflict.
          type: string
          readOnly: true
        asset:
          $ref: '#/components/schemas/Asset'
    AssetType:
      type: object
      anyOf:
        - $ref: '#/components/schemas/VMInfo'
        - $ref: '#/components/schemas/PodInfo'
        - $ref: '#/components/schemas/DirInfo'
        - $ref: '#/components/schemas/ContainerImageInfo'
        - $ref: '#/components/schemas/ContainerInfo'
      discriminator:
        propertyName: objectType
        mapping:
          VMInfo: '#/components/schemas/VMInfo'
          PodInfo: '#/components/schemas/PodInfo'
          DirInfo: '#/components/schemas/DirInfo'
          ContainerImageInfo: '#/components/schemas/ContainerImageInfo'
          ContainerInfo: '#/components/schemas/ContainerInfo'
    VMInfo:
      type: object
      properties:
        objectType:
          type: string
        instanceID:
          type: string
        instanceProvider:
          $ref: '#/components/schemas/CloudProvider'
        location:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
        securityGroups:
          type: array
          items:
            $ref: '#/components/schemas/SecurityGroup'
          nullable: true
        image:
          type: string
        instanceType:
          type: string
        platform:
          type: string
        launchTime:
          type: string
          format: date-time
        rootVolume:
          $ref: '#/components/schemas/RootVolume'
      required:
        - objectType
        - instanceID
        - location
        - instanceType
        - image
        - platform
        - launchTime
        - rootVolume
    PodInfo:
      type: object
      properties:
        objectType:
          type: string
        podName:
          type: string
        location:
          type: string
      required:
        - objectType
    ContainerInfo:
      type: object
      properties:
        objectType:
          type: string
        containerName:
          type: string
        location:
          type: string
        image:
          $ref: '#/components/schemas/ContainerImageInfo' 
        containerID:
          type: string
        createdAt:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
      required:
        - objectType
        - containerID
    ContainerImageInfo:
      type: object
      properties:
        objectType:
          type: string
        imageID:
          type: string
        repoTags:
          type: array
          items:
            type: string
        repoDigests:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
        architecture:
          type: string
        os:
          type: string
        size:
          type: integer
          format: int64
      required:
        - objectType
        - imageID
    DirInfo:
      type: object
      properties:
        objectType:
          type: string
        dirName:
          type: string
        location:
          type: string
      required:
        - objectType
...
    ProviderRelationship:
      type: object
      description: Describes a relationship to a provider which can be expanded.
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            revision:
              type: integer
              readOnly: true
            displayName:
              description: Human-readable name for the provider.
              type: string
              readOnly: true
            lastHeartbeatTime:
              description: Timestamp of the last heartbeat from the provider.
              type: string
              format: date-time
              readOnly: true
            status:
              $ref: '#/components/schemas/ProviderStatus'
              readOnly: true
            providerRuntimeVersion:
              description: Version of the provider runtime that the provider was compiled with.
              type: string
              readOnly: true
...
    ProviderStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the health state of the provider.

            | State     | Description                                    |
            | --------- | ---------------------------------------------- |
            | Healthy   | Provider is healthy.                           |
            | Unhealthy | Heartbeat received, but Provider is unhealthy. |
            | Unknown   | No heartbeat received from Provider.           |
          type: string
          enum:
            - Healthy
            - Unhealthy
            - Unknown
        reason:
          description: Machine-readable, UpperCamelCase text indicating the reason for the condition's last transition.
          type: string
          enum:
            - HeartbeatReceived
            - NoHeartbeatReceived
        message:
          description: Human-readable message indicating details about the last state transition.
          type: string
        lastTransitionTime:
          description: Last date-time when the status has changed.
          type: string
          format: date-time
...
  responses:
    Success:
      description: Success message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  parameters:
    assetID:
      name: assetID
      in: path
      required: true
      schema:
        type: string
    odataFilter:
      name: "$filter"
      in: query
      schema:
        type: string
    odataSelect:
      name: "$select"
      in: query
      schema:
        type: string
    odataCount:
      name: "$count"
      in: query
      schema:
        type: boolean
    odataTop:
      name: "$top"
      in: query
      schema:
        type: integer
    odataSkip:
      name: "$skip"
      in: query
      schema:
        type: integer
    odataExpand:
      name: "$expand"
      in: query
      schema:
        type: string
    odataOrderBy:
      name: "$orderby"
      in: query
      schema:
        type: string
      x-go-name: "OrderBy"
...
    ifmatch:
      name: If-Match
      in: header
      schema:
        type: integer
...